Feature/notice

See merge request s06-webmobile1-sub2/S06P12A607!144

+++ backend/src/main/java/com/dd/api/service/AwsS3ServiceImpl.java
@@ -3,7 +3,6 @@ package com.dd.api.service;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URLEncoder;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -29,10 +28,13 @@ import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import com.amazonaws.util.IOUtils;
 import com.dd.api.dto.response.FilesResponseDto;
+import com.dd.db.entity.board.Notice;
 import com.dd.db.entity.files.Files;
+import com.dd.db.entity.files.NoticeFile;
 import com.dd.db.entity.files.ProfileImg;
 import com.dd.db.entity.user.User;
 import com.dd.db.repository.FileRepository;
+import com.dd.db.repository.NoticeFileRepository;
 import com.dd.db.repository.ProfileImgRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -48,9 +50,9 @@ public class AwsS3ServiceImpl implements AwsS3Service {
     
     private final AmazonS3Client amazonS3Client;
 	
-	private final FileRepository fileRepository;
-	
 	private final ProfileImgRepository profileImgRepository;
+
+	private final NoticeFileRepository noticeFileRepository;
 
 	private final JwtTokenService jwtTokenService;
     
@@ -58,6 +60,7 @@ public class AwsS3ServiceImpl implements AwsS3Service {
 	@Override
 	public List<String> uploadFile(String accessToken, List<MultipartFile> multipartFile) {
 		User user = jwtTokenService.convertTokenToUser(accessToken);
+		
 		List<String> fileNameList = new ArrayList<>();
 		
 		multipartFile.forEach(file -> {
@@ -73,30 +76,88 @@ public class AwsS3ServiceImpl implements AwsS3Service {
 	                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
 	            }
 	            
-	            Files fileEntity = Files.builder()
+	            fileNameList.add(fileName);
+	        });
+		return fileNameList;
+	}
+	
+	@Transactional
+	@Override
+	public List<String> uploadFile(User user, Notice notice, List<MultipartFile> multipartFile) {
+		List<String> fileNameList = new ArrayList<>();
+		
+		multipartFile.forEach(file -> {
+				System.out.println(file.getOriginalFilename());
+	            String fileName = createFileName(file.getOriginalFilename());
+	            ObjectMetadata objectMetadata = new ObjectMetadata();
+	            objectMetadata.setContentLength(file.getSize());
+	            objectMetadata.setContentType(file.getContentType());
+
+	            try(InputStream inputStream = file.getInputStream()) {
+	                amazonS3.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
+	                        .withCannedAcl(CannedAccessControlList.PublicRead));
+	            } catch(IOException e) {
+	                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
+	            }
+	            
+	            NoticeFile noticefile = NoticeFile.builder()
 	            		.newFileName(fileName)
 	            		.originFileName(file.getOriginalFilename())
 	            		.user(user)
+	            		.notice(notice)
 	            		.build();
-	            
-	            fileRepository.save(fileEntity);
+
+	            noticeFileRepository.save(noticefile);
 
 	            fileNameList.add(fileName);
 	        });
 		return fileNameList;
 	}
+	
+	@Transactional
+	@Override
+	public String uploadFile(User user, Notice notice, MultipartFile file) {
+		
+	            String fileName = createFileName(file.getOriginalFilename());
+	            ObjectMetadata objectMetadata = new ObjectMetadata();
+	            objectMetadata.setContentLength(file.getSize());
+	            objectMetadata.setContentType(file.getContentType());
+
+	            try(InputStream inputStream = file.getInputStream()) {
+	                amazonS3.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
+	                        .withCannedAcl(CannedAccessControlList.PublicRead));
+	            } catch(IOException e) {
+	                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
+	            }
+	            
+	            NoticeFile noticefile = NoticeFile.builder()
+	            		.newFileName(fileName)
+	            		.originFileName(file.getOriginalFilename())
+	            		.user(user)
+	            		.notice(notice)
+	            		.build();
+
+	            noticeFileRepository.save(noticefile);
+
+		return fileName;
+	}
 
 	@Transactional
 	@Override
-	public void deleteFile(String accessToken, String fileName) {
+	public void deleteNoticeFile(Notice notice) {
+		NoticeFile noticeFile = noticeFileRepository.findByNotice(notice).orElse(null);
+		if(noticeFile == null)
+			return;
+		
+		String fileName = noticeFile.getNewFileName();
 		amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
-		Files deleteFile = fileRepository.findByNewFileName(fileName).get();
-		fileRepository.delete(deleteFile);
+		noticeFileRepository.delete(noticeFile);
+//		Files deleteFile = fileRepository.findByNewFileName(fileName).get();
+//		fileRepository.delete(deleteFile);
 	}
-
+	
 	@Override
 	public String createFileName(String fileName) {
-//		return String.valueOf(LocalDateTime.now()).concat(fileName);
 		return UUID.randomUUID().toString().concat(getFileExtension(fileName));
 	}
 
@@ -161,9 +222,9 @@ public class AwsS3ServiceImpl implements AwsS3Service {
 		    		.user(user)
 		    		.build();
         } else {
+//        	amazonS3.deleteObject(new DeleteObjectRequest(bucket, profileImg.getNewFileName()));
         	profileImg.updateImg(multipartFile.getOriginalFilename(), fileName);
         }
-        
         profileImgRepository.save(profileImg);
         
 		return fileName;
